+++ lib/marhan_cli/commands/network.rb
@@ -51,14 +51,5 @@ module MarhanCli
       options[option_name]
     end
 
-    def get_or_ask(option_name)
-      options[option_name] || ask_for_option(:host)
-    end
-
-    def ask_for_option(option)
-      value = ask("Please enter #{option}:")
-      raise Thor::Error, "You must enter a value for that field." if value.empty?
-      value
-    end
   end
 end